Chatter refactor

+++ dbactions/qdbTable.js
@@ -23,12 +23,14 @@ exports.init = (db, callback) => {
     score TEXT,
     votes TEXT,
     attachment BLOB,
-    attachmentUrl TEXT)
+    attachmentUrl TEXT,
+    guild TEXT NOT NULL)
     `);
   callback();
 };
 
-exports.get = (id, callback) => {
+exports.get = (id, guild, callback) => {
+  const guildId = guild.id ? guild.id : guild;
   let db = new sqlite3.Database('goosedb.sqlite', (err) => {
     if (err) {
       return console.error(err.message);
@@ -36,17 +38,14 @@ exports.get = (id, callback) => {
   });
   const idInt = Number(id);
   if(!Number.isNaN(idInt)) {
-    db.all('SELECT * FROM qdb WHERE id = $id',{$id:idInt}, callback).close(onClose);
+    db.all('SELECT * FROM qdb WHERE id = $id AND guild = $guild',{$id:idInt, $guild:guildId}, callback).close(onClose);
   } else {
-    db.all('SELECT * FROM qdb', callback).close((err) => {
-      if (err) {
-        return console.error(err.message);
-      }
-    });
+    db.all('SELECT * FROM qdb', callback).close(onClose);
   }
 };
 
-exports.like = (like, callback) => {
+exports.like = (like, guild, callback) => {
+  const guildId = guild.id ? guild.id : guild;
   let db = new sqlite3.Database('goosedb.sqlite', (err) => {
     if (err) {
       return console.error(err.message);
@@ -56,7 +55,7 @@ exports.like = (like, callback) => {
     return [];
   }
   const wildLike = `%${like}%`;
-  db.all('SELECT * FROM qdb WHERE body LIKE $like', {$like:wildLike}, callback).close(onClose);
+  db.all('SELECT * FROM qdb WHERE body LIKE $like AND guild = $guild', {$like:wildLike, $guild:guildId}, callback).close(onClose);
 };
 
 exports.delete = (qid, author, callback) => {
@@ -65,11 +64,10 @@ exports.delete = (qid, author, callback) => {
       return console.error(err.message);
     }
   });
-  // delete from qdb where id=17 and author_id="<@166052926777851904>"
-  db.run('DELETE FROM qdb WHERE id=$id and author_id=$author', {$id:qid, $author:author}, callback).close(onClose);
+  db.run('DELETE FROM qdb WHERE id=$id and (author_id=$author OR author_id is null)', {$id:qid, $author:`${author}`}, callback).close(onClose);
 };
 
-exports.vote= (qid, score, votes, callback) => {
+exports.vote = (qid, score, votes, callback) => {
   let db = new sqlite3.Database('goosedb.sqlite', (err) => {
     if (err) {
       return console.error(err.message);
@@ -79,7 +77,8 @@ exports.vote= (qid, score, votes, callback) => {
   db.run('UPDATE qdb set score=$score, votes=$votes WHERE id=$id', {$score:score, $votes:votes, $id:qid}, callback).close(onClose);
 };
 
-exports.add = (newQuote, message, callback, attachmentUrl, blob = Buffer.from([]).toString('base64')) => {
+exports.add = (newQuote, message, callback, options) => {
+  const {attachmentUrl, blob  = Buffer.from([]).toString('base64'), tags, notes} = options;
   let db = new sqlite3.Database('goosedb.sqlite', (err) => {
     if (err) {
       return console.error(err.message);
@@ -89,19 +88,22 @@ exports.add = (newQuote, message, callback, attachmentUrl, blob = Buffer.from([]
   // const tags = '';
   const score = '0';
   const votes = '0';
+  const messageUserId = message.user ? `${message.user}` : message.author.id;
+  const guildId = message.guild.id;
   db.run('INSERT INTO qdb' +
-    ' (body, created, author_id, score, votes, attachment, attachmentUrl)' +
-    ' VALUES ($body, $created, $author_id, $score, $votes, $blob, $au)', { $body: newQuote, $created: Date(), $author_id: message.author, $score: score, $votes: votes, $blob:blob, $au:attachmentUrl }, function (err) {
+    ' (body, notes, tags, created, author_id, guild, score, votes, attachment, attachmentUrl)' +
+    ' VALUES ($body, $notes, $tags, $created, $author_id, $guild, $score, $votes, $blob, $au)', { $body: newQuote, $notes:notes, $tags:tags, $created: Date(), $author_id: messageUserId, $guild: guildId, $score: score, $votes: votes, $blob:blob, $au:attachmentUrl }, function (err) {
     if (err) {
       return console.log(err.message);
     }
     // get the last insert id
     console.log(`A row has been inserted with rowid ${this.lastID}`);
-    callback(this);
+    callback(this, guildId);
   }).close(onClose);
 };
 
-exports.load = (filename = 'dbactions/qdb.json') => {
+exports.load = (filename = 'dbactions/qdb.json', guild='12345678987654321') => {
+  const guildId = guild.id ? guild.id : guild;
   let db = new sqlite3.Database('goosedb.sqlite', (err) => {
     if (err) {
       return console.error(err.message);
@@ -120,9 +122,9 @@ exports.load = (filename = 'dbactions/qdb.json') => {
         data.forEach(row => {
           const { body, notes, tags, created, status, deleted, author_id, author_ip, score, votes } = row;
           db.run('INSERT INTO qdb' +
-          ' (body, notes, tags, created, status, deleted, author_id, author_ip, score, votes)'+
+          ' (body, notes, tags, created, status, deleted, author_id, author_ip, guild, score, votes)'+
           ' VALUES ($body, $notes, $tags, $created, $status, $deleted, $author_id, $author_ip, $score, $votes)',
-          {$body:body, $notes:notes, $tags:tags, $created:created, $status: status, $deleted:deleted, $author_id:author_id, $author_ip:author_ip, $score:score, $votes:votes}, function (err) {
+          {$body:body, $notes:notes, $tags:tags, $created:created, $status: status, $deleted:deleted, $author_id:author_id, $author_ip:author_ip, $guild:guildId, $score:score, $votes:votes}, function (err) {
             if (err) {
               return console.log(err.message);
             }
@@ -135,4 +137,3 @@ exports.load = (filename = 'dbactions/qdb.json') => {
     });
   });
 };
-
